backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
reportlab==4.1.0
google-generativeai==0.7.2
python-multipart==0.0.9
firebase-admin
sqlalchemy
psycopg2-binary
python-dotenv
fastapi-mail
PyMuPDF 
reportlab